requirements.txt
moviepy==1.0.3
openai_whisper==20230314
pytube==15.0.0
